requirements.txt
asgiref
asttokens
attrs
decorator
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
exceptiongroup
executing
inflection
ipython
jedi
jsonschema
jsonschema-specifications
matplotlib-inline
parso
pexpect
pillow
prompt-toolkit
psycopg2
ptyprocess
pure-eval
Pygments
PyJWT
python-dotenv
PyYAML
referencing
rpds-py
six
sqlparse
stack-data
traitlets
typing_extensions
uritemplate
wcwidth
